feat: add all ids check initalized
