+++ adcio_core/src/main/java/com/corcaai/adcio_core/error/NotInitializedException.kt
@@ -1,5 +1,5 @@
 package com.corcaai.adcio_core.error
 
-data class IsNeedClientIdException(
+data class NotInitializedException(
     override val message: String
 ) : RuntimeException()
